refactor(cluster-init): use cobra cmd
