+++ cmd/cluster-init/cmd/onboard/build_clusters.go
@@ -1,4 +1,4 @@
-package main
+package onboard
 
 import (
 	"os"
