+++ cmd/cluster-init/cmd/onboard/onboard_test.go
@@ -1,4 +1,4 @@
-package main
+package onboard
 
 import (
 	"errors"
