+++ cmd/cluster-init/cmd/onboard/testdata/clusters/build-clusters/existingCluster/common_cert_manager
