+++ cmd/cluster-init/cmd/onboard/testdata/clusters/build-clusters/existingCluster/common_except_app.ci
