+++ cmd/cluster-init/cmd/onboard/update_bootstrap_secrets_test.go
@@ -1,4 +1,4 @@
-package main
+package onboard
 
 import (
 	"errors"
