+++ cmd/cluster-init/cmd/onboard/update_prow_plugin_config.go
@@ -1,4 +1,4 @@
-package main
+package onboard
 
 import (
 	"fmt"
