+++ cmd/cluster-init/cmd/onboard/update_sanitize_prow_jobs_test.go
@@ -1,4 +1,4 @@
-package main
+package onboard
 
 import (
 	"testing"
