+++ cmd/cluster-init/cmd/onboard/update_secret_generator.go
@@ -1,4 +1,4 @@
-package main
+package onboard
 
 import (
 	"fmt"
