+++ cmd/cluster-init/cmd/onboard/update_secret_generator_test.go
@@ -1,4 +1,4 @@
-package main
+package onboard
 
 import (
 	"errors"
