+++ cmd/cluster-init/cmd/onboard/update_sync_rover_groups.go
@@ -1,4 +1,4 @@
-package main
+package onboard
 
 import (
 	"fmt"
